Store version files separately from version data.
When uploading several releases the version data almost never changes.

+++ server/devpi_server/model.py
@@ -22,6 +22,7 @@ from .filestore import FileEntry
 from .log import threadlog
 from .markers import unknown
 from .normalized import normalize_name
+from .readonly import ensure_deeply_readonly
 from .readonly import get_mutable_deepcopy
 from operator import iconcat
 from typing import TYPE_CHECKING
@@ -1302,6 +1303,7 @@ class PrivateStage(BaseStage):
         errors. """
         if self.customizer.readonly:
             raise ReadonlyIndex("index is marked read only")
+        metadata = {k: v for k, v in metadata.items() if not k.startswith("+")}
         # use a copy, as validate_metadata actually removes metadata_version
         validate_metadata(dict(metadata))
         self._set_versiondata(metadata)
@@ -1315,17 +1317,34 @@ class PrivateStage(BaseStage):
             user=self.username, index=self.index,
             project=normalize_name(project), version=version)
 
+    def key_versionfilelist(self, project, version):
+        return self.keyfs.VERSIONFILELIST(
+            user=self.username,
+            index=self.index,
+            project=normalize_name(project),
+            version=version,
+        )
+
+    def key_versionfile(self, project, version, filename):
+        return self.keyfs.VERSIONFILE(
+            user=self.username,
+            index=self.index,
+            project=normalize_name(project),
+            version=version,
+            filename=filename,
+        )
+
     def _set_versiondata(self, metadata):
+        assert "+elinks" not in metadata
         project = normalize_name(metadata["name"])
         version = metadata["version"]
         key_projversion = self.key_projversion(project, version)
         with key_projversion.update() as versiondata:
             versiondata.update(metadata)
         threadlog.info("set_metadata %s-%s", project, version)
-        versions = self.key_projversions(project).get_mutable()
-        if version not in versions:
-            versions.add(version)
-            self.key_projversions(project).set(versions)
+        with self.key_projversions(project).update() as versions:
+            if version not in versions:
+                versions.add(version)
         self.add_project_name(project)
 
     def add_project_name(self, project):
@@ -1360,6 +1379,7 @@ class PrivateStage(BaseStage):
         linkstore.remove_links()
         versions.remove(version)
         self.key_projversion(project, version).delete()
+        self.key_versionfilelist(project, version).delete()
         self.key_projversions(project).set(versions)
         if cleanup:
             if not versions:
@@ -1381,7 +1401,17 @@ class PrivateStage(BaseStage):
     def list_versions_perstage(self, project):
         return self.key_projversions(project).get()
 
-    def get_last_project_change_serial_perstage(self, project, at_serial=None):  # noqa: PLR0911
+    def _get_elinks(self, project, version):
+        filenames = self.key_versionfilelist(project, version).get()
+        return [
+            self.key_versionfile(project, version, filename).get()
+            for filename in sorted(filenames)
+        ]
+
+    def get_has_versiondata_perstage(self, project, version):
+        return self.key_projversion(project, version).exists()
+
+    def get_last_project_change_serial_perstage(self, project, at_serial=None):  # noqa: PLR0911, PLR0912
         project = normalize_name(project)
         tx = self.keyfs.tx
         if at_serial is None:
@@ -1410,24 +1440,59 @@ class PrivateStage(BaseStage):
             # the project never existed or was deleted and didn't have versions
             return -1
         (last_serial, versions) = info
+        if last_serial >= at_serial:
+            return last_serial
         if versions is None:
             # was deleted
             return last_serial
-        version = get_latest_version(versions)
-        info = tx.get_last_serial_and_value_at(
-            self.key_projversion(project, version),
-            at_serial,
-            raise_on_error=False,
-        )
-        if info is None:
-            # never existed
-            return -1
-        (version_serial, version) = info
-        return max(last_serial, version_serial)
+        for version in versions:
+            info = tx.get_last_serial_and_value_at(
+                self.key_projversion(project, version), at_serial, raise_on_error=False
+            )
+            if info is None:
+                continue
+            (version_serial, version_info) = info
+            last_serial = max(last_serial, version_serial)
+            if last_serial >= at_serial:
+                return last_serial
+            info = tx.get_last_serial_and_value_at(
+                self.key_versionfilelist(project, version),
+                at_serial,
+                raise_on_error=False,
+            )
+            if info is None:
+                continue
+            (versionfiles_serial, versionfiles_info) = info
+            last_serial = max(last_serial, versionfiles_serial)
+            if last_serial >= at_serial:
+                return last_serial
+            for filename in versionfiles_info:
+                info = tx.get_last_serial_and_value_at(
+                    self.key_versionfile(project, version, filename),
+                    at_serial,
+                    raise_on_error=False,
+                )
+                if info is None:
+                    continue
+                (versionfile_serial, versionfile_info) = info
+                last_serial = max(last_serial, versionfile_serial)
+                if last_serial >= at_serial:
+                    return last_serial
+                last_serial = max(last_serial, versionfile_serial)
+                if last_serial >= at_serial:
+                    return last_serial
+        return last_serial
 
-    def get_versiondata_perstage(self, project, version):
+    def get_versiondata_perstage(self, project, version, *, with_elinks=True):
         project = normalize_name(project)
-        return self.key_projversion(project, version).get()
+        verdata = self.key_projversion(project, version).get()
+        assert "+elinks" not in verdata
+        if with_elinks:
+            elinks = self._get_elinks(project, version)
+            if elinks:
+                verdata = get_mutable_deepcopy(verdata)
+                verdata["+elinks"] = elinks
+        return ensure_deeply_readonly(verdata)
 
     def get_simplelinks_perstage(self, project):
         data = self.key_projsimplelinks(project).get()
@@ -1445,8 +1510,9 @@ class PrivateStage(BaseStage):
             linkstore = self.get_linkstore_perstage(project, version)
             releases = linkstore.get_links(Rel.ReleaseFile)
             links.extend(map(make_key_and_href, releases))
-            require_python = self.get_versiondata_perstage(project,
-                    version).get('requires_python')
+            require_python = self.get_versiondata_perstage(
+                project, version, with_elinks=False
+            ).get("requires_python")
             requires_python.extend([require_python] * len(releases))
         data_dict = {u"links":links, u"requires_python":requires_python}
         self.key_projsimplelinks(project).set(data_dict)
@@ -1464,12 +1530,12 @@ class PrivateStage(BaseStage):
             raise ReadonlyIndex("index is marked read only")
         project = normalize_name(project)
         filename = ensure_unicode(filename)
-        if not self.get_versiondata_perstage(project, version):
+        if not self.get_has_versiondata_perstage(project, version):
             # There's a chance the version was guessed from the
             # filename, which might have swapped dashes to underscores
             if '_' in version:
                 version = version.replace('_', '-')
-                if not self.get_versiondata_perstage(project, version):
+                if not self.get_has_versiondata_perstage(project, version):
                     raise MissesRegistration("%s-%s", project, version)
             else:
                 raise MissesRegistration("%s-%s", project, version)
@@ -1499,7 +1565,7 @@ class PrivateStage(BaseStage):
             threadlog.info("store_doczip: derived version of %s is %s",
                            project, version)
         basename = "%s-%s.doc.zip" % (project, version)
-        if not self.get_versiondata_perstage(project, version):
+        if not self.get_has_versiondata_perstage(project, version):
             self.set_versiondata({'name': project, 'version': version})
         linkstore = self.get_mutable_linkstore_perstage(project, version)
         return linkstore.create_linked_entry(
@@ -1514,11 +1580,17 @@ class PrivateStage(BaseStage):
         """ get link of documentation zip or None if no docs exists. """
         linkstore = self.get_linkstore_perstage(project, version)
         links = linkstore.get_links(rel=Rel.DocZip)
-        if len(links) > 1:
-            threadlog.warning(
-                "Multiple documentation files for %s-%s, returning newest",
-                project, version)
-        return links[-1] if links else None
+        link = links[-1] if links else None
+        if len(links) > 1 and link is not None:
+            # the order is not defined, but since devpi-server 2.4.0 it
+            # shouldn't be possible to get into this case anyway
+            threadlog.warn(
+                "Multiple documentation files for %s-%s, returning %s",
+                project,
+                version,
+                link.entrypath,
+            )
+        return link
 
     def get_doczip_entry(self, project, version):
         """ get entry of documentation zip or None if no docs exists. """
@@ -1555,6 +1627,30 @@ class PrivateStage(BaseStage):
                 last_serial = max(last_serial, version_serial)
                 if last_serial >= at_serial:
                     return last_serial
+                try:
+                    (versionfiles_serial, versionfilenames) = (
+                        tx.get_last_serial_and_value_at(
+                            self.key_versionfilelist(project, version), at_serial
+                        )
+                    )
+                except KeyError:
+                    continue
+                last_serial = max(last_serial, versionfiles_serial)
+                if last_serial >= at_serial:
+                    return last_serial
+                for filename in versionfilenames:
+                    try:
+                        (versionfile_serial, versionfile_info) = (
+                            tx.get_last_serial_and_value_at(
+                                self.key_versionfile(project, version, filename),
+                                at_serial,
+                            )
+                        )
+                    except KeyError:
+                        continue
+                    last_serial = max(last_serial, versionfile_serial)
+                    if last_serial >= at_serial:
+                        return last_serial
         # no project uploaded yet
         index_key = self.key_index
         (index_serial, index_config) = tx.get_last_serial_and_value_at(
@@ -1689,15 +1785,37 @@ class LinkStore:
     def __repr__(self):
         return f"<{self.__class__.__name__} {self.project} {self.stage.name} {self.version}>"
 
+    def get_links(self, rel=None, basename=None, entrypath=None, for_entrypath=None):
+        if isinstance(for_entrypath, ELink):
+            for_entrypath = for_entrypath.relpath
+        elif for_entrypath is not None:
+            assert "#" not in for_entrypath
+
+        def fil(link):
+            return (
+                (not rel or rel == link.rel)
+                and (not basename or basename == link.basename)
+                and (not entrypath or entrypath in (link.entrypath, link.relpath))
+                and (not for_entrypath or for_entrypath == link.for_entrypath)
+            )
+
+        return [
+            elink
+            for linkdict in self.verdata.get("+elinks", [])
+            if fil(elink := ELink(self.filestore, linkdict, self.project, self.version))
+        ]
+
     @property
     def metadata(self):
-        return {
-            k: v
-            for k, v in get_mutable_deepcopy(self.verdata).items()
-            if not k.startswith("+")}
+        return ensure_deeply_readonly(
+            {k: v for k, v in self.verdata.items() if not k.startswith("+")}
+        )
 
-    def get_file_entry(self, relpath):
-        return self.filestore.get_file_entry(relpath)
+
+class MutableLinkStore(LinkStore):
+    def __init__(self, stage, project, version):
+        super().__init__(stage, project, version)
+        self.verdata = get_mutable_deepcopy(self.verdata)
 
     def create_linked_entry(
         self, rel, basename, content_or_file, *, hashes, last_modified=None
@@ -1751,42 +1869,28 @@ class LinkStore:
         return self._add_link_to_file_entry(rel, entry, for_link=for_link)
 
     def remove_links(self, rel=None, basename=None, for_entrypath=None):
-        linkdicts = self._get_inplace_linkdicts()
         del_links = self.get_links(rel=rel, basename=basename, for_entrypath=for_entrypath)
         was_deleted = []
-        for link in del_links:
-            link.entry.delete()
-            linkdicts.remove(link.linkdict)
-            was_deleted.append(link.relpath)
-            threadlog.info("deleted %r link %s", link.rel, link.relpath)
-        if linkdicts:
+        if del_links:
+            with self.stage.key_versionfilelist(
+                self.project, self.version
+            ).update() as versionfilenames:
+                for link in del_links:
+                    filename = link.entry.basename
+                    self.verdata["+elinks"].remove(link.linkdict)
+                    link.entry.delete()
+                    versionfilenames.remove(filename)
+                    self.stage.key_versionfile(
+                        self.project, self.version, filename
+                    ).delete()
+                    was_deleted.append(link.relpath)
+                    threadlog.info("deleted %r link %s", link.rel, link.relpath)
+        if (
+            was_deleted
+            and self.stage.key_versionfilelist(self.project, self.version).get()
+        ):
             for relpath in was_deleted:
                 self.remove_links(for_entrypath=relpath)
-        if was_deleted:
-            self._mark_dirty()
-
-    def get_links(
-        self,
-        rel: Rel | None = None,
-        basename: str | None = None,
-        entrypath: str | None = None,
-        for_entrypath: ELink | str | None = None,
-    ) -> list[ELink]:
-        if isinstance(for_entrypath, ELink):
-            for_entrypath = for_entrypath.relpath
-        elif for_entrypath is not None:
-            assert "#" not in for_entrypath
-
-        def fil(link):
-            return (
-                (not rel or rel == link.rel)
-                and (not basename or basename == link.basename)
-                and (not entrypath or entrypath in (link.entrypath, link.relpath))
-                and (not for_entrypath or for_entrypath == link.for_entrypath)
-            )
-
-        return list(filter(fil, [ELink(self.filestore, linkdict, self.project, self.version)
-                           for linkdict in self.verdata.get("+elinks", [])]))
 
     def _create_file_entry(
         self, basename, content_or_file, *, hashes, ref_hash_spec=None
@@ -1803,12 +1907,6 @@ class LinkStore:
         entry.version = self.version
         return entry
 
-    def _mark_dirty(self):
-        self.stage._set_versiondata(self.verdata)
-
-    def _get_inplace_linkdicts(self):
-        return self.verdata.setdefault("+elinks", [])
-
     def _add_link_to_file_entry(self, rel, file_entry, for_link=None):
         new_linkdict = {
             "rel": str(rel),
@@ -1819,17 +1917,17 @@ class LinkStore:
         if for_link:
             assert isinstance(for_link, ELink)
             new_linkdict["for_entrypath"] = for_link.relpath
-        linkdicts = self._get_inplace_linkdicts()
-        linkdicts.append(new_linkdict)
+        with self.stage.key_versionfilelist(
+            self.project, self.version
+        ).update() as versionfilelist:
+            assert file_entry.basename not in versionfilelist, file_entry.basename
+            versionfilelist.add(file_entry.basename)
+        self.stage.key_versionfile(self.project, self.version, file_entry.basename).set(
+            new_linkdict
+        )
+        self.verdata.setdefault("+elinks", []).append(new_linkdict)
         threadlog.info("added %r link %s", rel, file_entry.relpath)
-        self._mark_dirty()
         return ELink(self.filestore, new_linkdict, self.project, self.version)
-
-
-class MutableLinkStore(LinkStore):
-    def __init__(self, stage, project, version):
-        super().__init__(stage, project, version)
-        self.verdata = get_mutable_deepcopy(self.verdata)
 
 
 @total_ordering
@@ -2001,6 +2099,8 @@ def add_keys(xom, keyfs):
     keyfs.add_key("PROJVERSIONS", "{user}/{index}/{project}/.versions", set)
     keyfs.add_key("PROJVERSION", "{user}/{index}/{project}/{version}/.config", dict)
     keyfs.add_key("PROJNAMES", "{user}/{index}/.projects", set)
+    keyfs.add_key("VERSIONFILELIST", "{user}/{index}/{project}/{version}/.files", set)
+    keyfs.add_key("VERSIONFILE", "{user}/{index}/{project}/{version}/{filename}", dict)
     keyfs.add_key("STAGEFILE",
                   "{user}/{index}/+f/{hashdir_a}/{hashdir_b}/{filename}", dict)
